ov_sample/python_samples/requirements.txt
matplotlib==3.1.3
scikit-build==0.11.1
requests==2.23.0
six==1.12.0
gym==0.17.3
opencv-python==4.4.0.44
pillow==8.2.0
scipy==1.5.4 
stable-baselines3==0.10.0 
tensorboard==2.4.0 
tensorboard-plugin-wit==1.7.0 
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.0+cu110
torchvision==0.8.1+cu110
torchaudio==0.7.0
